(refactor) se separaron las vistas
