+++ chatbot/urls.py
@@ -1,12 +1,12 @@
 from django.urls import path
-from . import views
+from .views import chat_views, conversation_views, feedback_views
 
 app_name = 'chatbot'
 
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('api/chat/', views.chat, name='chat_api'),
-    path('api/feedback/', views.feedback, name='feedback_api'),
-    path('api/conversations/', views.conversations, name='conversations_api'),
-    path('api/conversations/<uuid:conversation_id>/', views.get_conversation, name='get_conversation_api'),
+    path('', chat_views.index, name='index'),
+    path('api/chat/', chat_views.chat, name='chat_api'),
+    path('api/feedback/', feedback_views.feedback, name='feedback_api'),
+    path('api/conversations/', conversation_views.conversations, name='conversations_api'),
+    path('api/conversations/<uuid:conversation_id>/', conversation_views.get_conversation, name='get_conversation_api'),
 ]
