+++ chatbot/views/__init__.py
@@ -1,0 +1,12 @@
+from .chat_views import index, chat, init_chat_service
+from .conversation_views import conversations, get_conversation
+from .feedback_views import feedback
+
+__all__ = [
+    'index',
+    'chat',
+    'init_chat_service',
+    'conversations',
+    'get_conversation',
+    'feedback'
+]
